# src/app/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
...
  controller.CatItem:
    properties:
      adoption_price:
        type: integer
      cat_thumbnail_path:
        type: string
      contract_benefit:
        type: integer
      contract_days:
        type: integer
      deposit:
        type: integer
      id:
        type: integer
      level:
        type: string
      name:
        type: string
      pet_coin:
        type: integer
      price:
        type: integer
      reservation_price:
        type: integer
      status:
        type: integer
    type: object
...
  controller.GetCatListResponse:
    properties:
      cats:
        items:
          $ref: '#/definitions/controller.CatItem'
        type: array
      lower:
        type: integer
      total:
        type: integer
      upper:
        type: integer
    type: object
...
  controller.GetTimePeriodResponse:
    properties:
      lower:
        type: integer
      time_periods:
        items:
          $ref: '#/definitions/controller.TimePeriodItem'
        type: array
      total:
        type: integer
      upper:
        type: integer
    type: object
...
  controller.Message:
    properties:
      message:
        example: message
        type: string
    type: object
  controller.TimePeriodItem:
    properties:
      cats:
        items:
          $ref: '#/definitions/controller.CatItem'
        type: array
      end_at:
        type: string
      id:
        type: integer
      start_at:
        type: string
    type: object
...
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
host: '{{.Host}}'
info:
  contact: {}
  description: '{{.Description}}'
  license: {}
  title: '{{.Title}}'
  version: '{{.Version}}'
paths:
...
  /api/v1/certificate/transfer/{transferId}:
    get:
      consumes:
      - application/json
      description: get transfer cat certificate thumbnail from the database
      parameters:
      - description: transfer的ID
        in: path
        name: transferId
        required: true
        type: integer
      produces:
      - image/jpeg
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
  /api/v1/time/schedule:
    post:
      consumes:
      - multipart/form-data
      description: Add a new timePeriod
      parameters:
      - description: 起始時間(yyyy-MM-dd'T'HH:mm:ss)
        in: formData
        name: start_at
        required: true
        type: string
      - description: 結束時間(yyyy-MM-dd'T'HH:mm:ss)
        in: formData
        name: end_at
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Message'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
  /api/v1/time/schedules:
    get:
      consumes:
      - application/json
      description: get timePeriod list with date
      parameters:
      - description: 時段列表的lower
        in: query
        name: lower
        required: true
        type: integer
      - description: 時段列表的upper
        in: query
        name: upper
        required: true
        type: integer
      - description: 時段列表的日期(yyyy-MM-dd)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetTimePeriodResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
...
  /api/v1/transfer/{transferId}/certificate:
    post:
      consumes:
      - multipart/form-data
      description: post transfer cat certificate to database
      parameters:
      - description: transfer的ID
        in: path
        name: transferId
        required: true
        type: integer
      - description: certificate的縮圖檔案
        in: formData
        name: certificate
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Message'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
  /api/v1/transfer/{transferId}/migrate:
    post:
      consumes:
      - application/json
      description: post transfer cat migrate to the database
      parameters:
      - description: transfer的ID
        in: path
        name: transferId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
  /api/v1/transfer/cats:
    get:
      consumes:
      - application/json
      description: get transfer cat list with status from database
      parameters:
      - description: 貓列表的lower
        in: query
        name: lower
        required: true
        type: integer
      - description: 貓列表的upper
        in: query
        name: upper
        required: true
        type: integer
      - description: '貓列表的交易狀態( 待交易: 1 / 買家已上傳憑證 : 2/已完成 :3 /已取消)'
        in: query
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.GetCatListResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
            type: object
...
swagger: "2.0"